Move DBStorage class to own module
